Export & Auto Trans

+++ src/model/VocabCite.php
@@ -14,15 +14,15 @@ class VocabCite extends Model
     {
         $connection = config('trans-helper.database.connection') ?: config('database.default');
         $this->setConnection($connection);
-        $this->setTable(config('trans-helper.database.table.cite'));
+        $this->setTable('_tans_helper_cites');
         parent::__construct($attributes);
     }
 
     public function terms()
     {
         return $this->belongsToMany(
-            config('trans-helper.model.term'),
-            config('trans-helper.database.table.link'),
+            VocabTerm,
+            '_trans_helper_links',
             'cited',
             'vocab',
             'id',
@@ -32,26 +32,23 @@ class VocabCite extends Model
 
     public function sweep()
     {
-        $line = explode("\n", file_get_contents(base_path().$this->file));
+        $line = explode("\n", file_get_contents(base_path() . $this->file));
         $line = $line[$this->line - 1] ?? '';
+        $count = 0;
         if ($line) {
-            $terms = call_user_func_array(
-                'array_merge',
-                array_map(
-                    function ($u) {
-                        return ["localize('{$u->term}')", 'localize("'.$u->term.'")'];
-                    },
-                    $this->terms()->get()->all()
-                )
-            );
-            $matched = array_filter(array_map(function ($u) use ($line) {
-                return strpos($line, $u);
-            }, $terms), 'strlen');
-
-            if (!empty($matched)) {
-                return;
+            foreach ($this->terms()->get() as $term) {
+                $keywords = ["localize('{$term->term}')", 'localize("' . $term->term . '")'];
+                $matched = array_filter(array_map(function ($u) use ($line) {
+                    return strpos($line, $u);
+                }, $keywords), 'strlen');
+                if (empty($matched)) {
+                    $this->terms()->detach($term->id);
+                }
+                $count += count($matched);
             }
         }
-        $this->delete();
+        if ($count == 0) {
+            $this->delete();
+        }
     }
 }
